Updating script to use Rdio API and redis storage
So, I got a little crazy with this. I've been looking for an excuse to
try out using redis for key:value storage and CSVs are, like, totally
lame.

Since this is basically just a first pass at this, I did all the fun
stuff in the url parsing loop. Like a crazy person. We'll refactor this
in later revisions.

The comments should explain most of what I changed, with TODOs
describing some of the things I want to add.