Updating script with more features
Here's the second pass at making this a useful script. I gave a lot more
thought to how I wanted to store/query the data this time, and
refactored _most_ of the code out of the rss parsing loop. Also, now
that we're storing the unavailable albums, we can query if they've
become available every time at the end of the script!

I added a key to store the value of the most recently parsed item in the
feed, so we don't even need to look at the whole feed every time. Cool.